Add QTorrent binary and move dots to library for reuse (#200)
* add qtorrent binary

* handle multi file torrents too

* cargo update

* cli improvements

* move format size and use u64

* improve rename

* optimize string clones

* output improvements

* move dots to library to allow for reuse

* tweak output

* add args to install script

* tweak directory naming and tweak indent

* fix filtering

* improve directory skip

* print index

* fix dot rename usage for files

* tweak print

* format internal name too

* print tweaks

* separate dot name formatting code

* update usage for qtorrent

* use DotFormat directly in QTorrent

* fix CI tests

+++ install.sh
@@ -8,6 +8,42 @@ DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 # shellcheck source=./common.sh
 source "$DIR/common.sh"
 
+USAGE="Usage: $0 [OPTIONS]
+
+Install all Rust binaries to the Cargo bin directory.
+
+OPTIONS: All options are optional
+    -f | --force
+        Force recompilation by touching source files.
+
+    --help
+        Display these instructions.
+
+    --verbose
+        Display commands being executed.
+"
+
+FORCE=false
+
+while [ $# -gt 0 ]; do
+    case "$1" in
+        -f | --force)
+            FORCE=true
+            ;;
+        --help)
+            echo "$USAGE"
+            exit 1
+            ;;
+        --verbose)
+            set -x
+            ;;
+        *)
+            print_error_and_exit "Unknown option: $1"
+            ;;
+    esac
+    shift
+done
+
 if [ -z "$(command -v cargo)" ]; then
     print_error_and_exit "Cargo not found in path. Maybe install rustup?"
 fi
@@ -15,16 +51,20 @@ fi
 print_magenta "Installing binaries..."
 cd "$REPO_ROOT"
 
-# Remove existing release binaries to force recompilation with current version number.
-if [ -d "target/release" ]; then
-    for executable in $(get_rust_executable_names); do
-        rm -f "target/release/${executable}"
-    done
-fi
+if [ "$FORCE" = true ]; then
+    print_yellow "Forcing recompilation..."
 
-# Touch source files to ensure recompilation
-touch src/lib.rs
-find src/bin -name "*.rs" -exec touch {} \;
+    # Remove existing release binaries to force recompilation with current version number.
+    if [ -d "target/release" ]; then
+        for executable in $(get_rust_executable_names); do
+            rm -f "target/release/${executable}"
+        done
+    fi
+
+    # Touch source files to ensure recompilation
+    touch src/lib.rs
+    find src/bin -name "*.rs" -exec touch {} \;
+fi
 
 cargo install --force --path "$REPO_ROOT"
 echo ""
